Move MultiSearchIndex into models
