+++ app/models/multi_search_index.rb
